Docker Links, URLs, and Ports as Parameters
Updated the bash scripts to pass information to the docker containers.
This includes the container link names and associated server URLs.
Make the java script configs able to read the env at runtime when
the services are being called.
Merged the server and approval data in the reservations service
before passing to UI

Change-Id: I1c2e00bee28cddfcaa3ba61d56a9d095185d32c3